# Conflicts:
#	project/common-impl/src/main/kotlin/ink/ptms/adyeshach/impl/entity/type/DefaultPainting.kt
#	project/common-impl/src/main/resources/core/description/entity_meta.desc

+++ project/compat-model-engine-v4/src/main/kotlin/ink/ptms/adyeshach/compat/modelengine4/DefaultModelEngine.kt
@@ -41,7 +41,7 @@ internal interface DefaultModelEngine : ModelEngine {
         if (isModelEngineHooked && modelEngineName.isNotBlank() && modelEngineUniqueId != null) {
             this as DefaultEntityInstance
             // 如果没有其他观察者，直接销毁模型而不是隐藏
-            if (!viewPlayers.hasVisiblePlayer()) {
+            if (!viewPlayers.hasVisiblePlayer() && modelEngineOptions?.isDestroyWhenInvisible == true) {
                 destroyModelEngine()
             } else {
                 getDummy()?.setForceViewing(viewer, false)
